+++ resources/aws_support_app_account_alias.rb
@@ -1,7 +1,6 @@
 # Import API specifics
 use "awscc_base"
 
-unified_mode true
 resource_name :aws_support_app_account_alias
 provides :aws_support_app_account_alias, target_mode: true, platform: "aws"